refactor: 유틸성 패키지 이름 변경 및 통합
