+++ backend/src/test/java/com/pickpick/message/ui/MessageControllerTest.java
@@ -11,7 +11,7 @@ import static org.springframework.restdocs.request.RequestDocumentation.requestP
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.pickpick.auth.support.JwtTokenProvider;
-import com.pickpick.config.RestDocsTestSupport;
+import com.pickpick.support.RestDocsTestSupport;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
