+++ backend/src/test/java/com/pickpick/slackevent/application/message/MessageDeletedServiceTest.java
@@ -1,15 +1,15 @@
 package com.pickpick.slackevent.application.message;
 
+import static com.pickpick.support.JsonUtils.toJson;
 import static org.assertj.core.api.Assertions.assertThat;
-import static utils.JsonUtils.toJson;
 
 import com.pickpick.channel.domain.Channel;
 import com.pickpick.channel.domain.ChannelRepository;
-import com.pickpick.config.DatabaseCleaner;
 import com.pickpick.member.domain.Member;
 import com.pickpick.member.domain.MemberRepository;
 import com.pickpick.message.domain.Message;
 import com.pickpick.message.domain.MessageRepository;
+import com.pickpick.support.DatabaseCleaner;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
