+++ backend/src/test/java/com/pickpick/support/RestDocsConfiguration.java
@@ -1,4 +1,4 @@
-package com.pickpick.config;
+package com.pickpick.support;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
