+++ backend/src/test/java/com/pickpick/support/RestDocsTestSupport.java
@@ -1,4 +1,4 @@
-package com.pickpick.config;
+package com.pickpick.support;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterEach;
